Move example integration tests into integration folder
