+++ tests/integration/mirage-auto-resolver.test.ts
@@ -1,12 +1,12 @@
 import {expect} from 'chai';
 import {buildSchema} from 'graphql';
 import defaultResolvers from './resolvers';
-import {patchWithAutoResolvers} from '../src/mirage/resolver-reducers/patch-with-auto';
+import {patchWithAutoResolvers} from '../../src/mirage/resolver-reducers/patch-with-auto';
 import {server as mirageServer} from './mirage';
 import defaultScenario from './mirage/scenarios/default';
 import {buildHandler, typeDefs} from './executable-schema';
-import {addContextToResolvers} from '../src/mirage/resolver-reducers/add-context';
-import resolversReduce from '../src/resolvers/reduce';
+import {addContextToResolvers} from '../../src/mirage/resolver-reducers/add-context';
+import resolversReduce from '../../src/resolvers/reduce';
 
 const schema = buildSchema(typeDefs);
