+++ tests/integration/mirage/models/post.ts
