+++ tests/integration/mirage/scenarios/default.ts
